New updates to my project
