+++ src/blocks/modal.css
@@ -26,10 +26,6 @@
   width: 358px;
 }
 
-/* .modal__avatar-form {
-  visibility: invisible;
-} */
-
 .modal__container {
   position: relative;
   background-color: #fff;
@@ -38,24 +34,6 @@
   flex-basis: min-content;
   max-width: 430px;
 }
-
-/* .modal__container_type_delete {
-  background-color: #fff;
-  visibility: visible;
-  max-width: 430px; 
-  height: 181px;
-  top: 100px;
-  border-radius: 10px;
-  opacity: 50%;
-}  */
-
-/*.modal__container_type_avatar {
-   width: 430px; 
-  height: 272px;
-  left: 425px;
-  top: 100px;
-  border: 1px;
-}  */
 
 .modal__close {
   border: none;
