+++ src/index.html
@@ -36,26 +36,6 @@
               type="button"
             ></button>
           </div>
-          <!-- <div> -->
-          <!-- <button class="avatar__edit-button"
-              id="avatar-edit-button"
-              type="button"> 
-            </button> -->
-          <!--<h3 class="modal__heading">Change profile picture</h3>
-            <form class="modal__form modal__avatar-form"> 
-              <span class="modal__error" id="avatar-input-error"></span>
-              <input
-                class="modal__input"
-                name="avatar"
-                type="URL"
-                id="avatar-input"
-                placeholder="Image URL"
-                required
-              />
-              <span class="modal__error" id="avatar-input-error"></span>
-              <button class="modal__button" type="submit">Save</button>
-            </form>-->
-          <!-- </div>   -->
           <div class="profile__info">
             <h1 class="profile__title"></h1>
             <button
